+++ CHANGES.md
@@ -17,6 +17,7 @@
 * Compiler/wasm: preserve physical identity of empty closures (#2207)
 * Compiler/wasm: fix crash when compiling some function calls (#2208)
 * Compiler: fix missing conditional simplification (#2217)
+* Compiler: fix Js_assign.simpl (#2218)
 
 
 # 6.3.2 (2026-02-15) - Lille